US110605 Implement getLocalizeResources
This is a Serge-friendly, dynamically-loaded, fallback-enabled i10n solution that allows us to generate complete translation sets on load. Currently, this is limited to just the `d2l-activity-assignment-editor` component.

Advantages of this approach are that we don't need a build step and can instead work directly with the JSON files that Serge works with. When the element is loaded, it will always start by loading the English langterms (on the assumption that this will be a complete set), after which it will load additional files if necessary. This loading-English-first approach is a super-safe option which could probably be skipped, assuming all of our langterm files contain the same keys. This approach has been tested to work with supported and unsupported localized and non-localized languages.

The primary disadvantage of this approach is that it is more work on page load - however, the work is a single network call for `lang="en"`, and for other supported languages as well if we skip the super-safe loading-English-first option.