US110605 Simplify getLocalizeResources
The `langs` that we're receiving from the `LocalizeMixin` have already done all the locale-prioritizing, so no need to duplicate it here. Instead, we can just fetch translations for each entry in `langs` (filtered to only ones we support), and build our fallback-capable map from those responses. This is hypothetically unbounded in requests, but practically speaking it's at most 3 or 4, and the responses are highly cacheable.