US110194 Add autosave component
When the assignment name is updated, this will show "Saving..." as the network request occurs. When the network request successfully completes, it shows ":heavy_check_mark: Saved" after a few moments (the delay is set by the component itself, and is not configurable). If the network request fails, the component hides itself (behaviour that is also defined by the component itself, and is not configurable).