US110605 Convert to mixin
At this point, we're looking to likely use a single `lang` folder for the entire ACE app, which means each component will need to extend the `ACELocalizeMixin`, which does the fetching and mapping of langterms. This mixin also now caches langterms, so that even with multiple elements on the page, we only do a request for a given language once per page load, after which it is cached.

This approach still doesn't completely answer all the questions, e.g. we're still sending the `context` Serge data down the wire, but at this point we're looking for something that works, but which doesn't inevitable cause us hassle down the road. Hopefully, handling translations will be a bit more developed in the near future, at which point we can extend/simplify/rework this approach.