US110194 Put d2l-save-status in editor, not detail view
There are going to be multiple input fields in the `d2l-activity-editor`, some of which will be within `d2l-activity-assignment-editor-detail`, but others which may not. Because of this, the save-status component will have to listen for changes from multiple children.

As it currently stands, the events that are generated by doing network fetches (`d2l-siren-entity-save-*`) do not bubble, which means we need to retarget them in the `editor-detail`. This seems preferable over making the events bubble, as otherwise we will end up in a circumstance where elements are going to be reacting to events coming from components that they don't care about on the page; better to be explicit and clear that we're just pushing this event up one level.