US110196 Fix _resolveUrl path
In #436 I moved the `d2l-activities-html-editor` component up one level, but didn't update this value. Strangely, changing it to just `../../` instead doesn't work, and is likely related to LitElement and `resolveUrl` not being friends. However, using `import.meta.url` does indeed work, and surprisingly also works for the demo! Confirmed that this works with a bundled and unbundled version of BSI.

This also removes the addition of `this.token`, which was thought to be required, but is not.