Add missing Serge config
Realized that while #431 is a better long-term solution, we'll need Serge to at least know that our repo is here in the meantime. This adds a config for the assignment activities editor. We could consider just making the activities editor one large Serge project, as there are likely many duplicate langterms between the different variants of the editor, but the downside to that approach is that we can't namespace langterms very nicely (as far as I'm aware), which means we'd end up with pretty huge langterm files. We can always change this in the future.