US110196 Fixup: HTML editor change event
This adds a more predictable `change` event to the `d2l-activity-html-editor`, which is internally hooked up to both the `input` and `change` event from `d2l-html-editor`. This ensures the outer component generates an event for all changes within the HTML editor, whereas the previous implementation wouldn't pick up e.g. inserting an image or changing the text format until the input text was also changed. This means that the parent component (`d2l-activity-assignment-editor-detail` in this case) only has to listen for this `change` event, and can use that to drive auto-saving the instructions value.

This also updates the size of the editor to be a minimum of 3 rows, which matches the design.