Add Serge config for generic components
In #440, a Serge config was added for the `assignmentActivityEditor`, but there will be several components within `d2l-activity-editor` that are not specific to assignments. To address this (based largely off [this thread](https://github.com/BrightspaceHypermediaComponents/activities/pull/446#pullrequestreview-306128414)), this adds a new config for the generic `activityEditor` langterms. There will likely be overlap in these terms, but this is the cleanest way to keep the domain-specific and domain-generic terms from cluttering each other up. This also means that eventually, when there are other specialized activity editors (e.g. `d2l-activity-survey-editor` or something), we'll have to add new Serge configs for them as well.

In order to support this, and use the same `getLocalizationResources` in all places, a minor refactor was required such that the method now takes in `import.meta.url` from any implementing class.