Ignore could-not-resolve-reference
This warning is causing `polymer lint` to fail on `save-status-mixin`, claiming it doesn't know what "class" is. This appears to be some weird bug in `polymer lint`, per https://github.com/Polymer/tools/issues/695 and https://github.com/Polymer/tools/issues/719. Adding an `import` to the mixin avoids this error for whatever reason - perhaps the linter is not recognizing the file correctly without an `import` at the beginning or something.