US110198 Add Submission Type select list
This adds a styled `select` element that presents the options for `submissionType` on the assignment. There are two caveats to this change:

1. The `completionType` dropdown is not included, which means the On Paper and Observed in Person submission types can't be used correctly (they will both just use "Automatically on due date" as the completion type for now)
2. The styles for the `select` have just been copy+pasted into a `select-styles` file, but are planned to be separated out somewhere common.

Both of these should be rectified soon - the first by another US, and the second once some further discussions occur with Gaudi et al.

This also includes some minor changes to not access `super._entity` directly for getters, which is somewhat bad form it turns out - instead, it's better to use local values that are set in the `_onAssignmentChange` handler.