Refactor SaveStatusMixin
The method `save` was pretty ambiguously-named, as it seemed to indicate that it was doing some sort of saving - in reality, though, it's just wrapping an `async` with the save-status-event-generating code. In addition, consumers probably shouldn't be accessing the single events themselves, so remove the methods to call them individually. (If there's a need to call them individually at some point, they can be added as mixin interface methods again.)