US111620 Add plaintext instructions editor
This passes down a user config setting to determine whether to use the HTML editor or a plaintext editor for the instructions. This could be represented in an API response somehow, but after some thought and trying to flesh that out, this was a much simpler solution. It's unfortunate that the interim components needs to pass down an attribute that they don't really care about (`htmlEditorEnabled`), but it's not all that much more code or complexity, and the presence of other text editing fields in the future would be easy to support.

The new `d2l-activity-text-editor` acts more-or-less the same as the HTML editor did, but now will display a `d2l-input-textarea` when the user doesn't want to use richtext editing.