Add Serge plugin for context
In the most recent translation run, we got back output that included translated `context` strings, which is not what we want. Turns out, this is due to the lack of this plugin, which translates the `context` into the expected description field for translation by Serge. This should mean that, after the next translation run, our translated files should no longer have the `context` strings in them at all, and instead be a fairly flat list of `langterm: "translation"` as expected.