Use Serge's parse_js parser
In the ongoing saga of a better integration between Serge and our components, this converts FACE to use [the `parse_js` parser](https://serge.io/docs/plugins/parser/parse_js/) in Serge. This addresses most of the requirements we had for improved i10n:

- Works with Serge
- Language files include context - for the JS parser, this is provided in the form of a comment after the langterm
- No build step - Serge takes the ES6 modules as inputs, and produces ES6 modules as output
- Only change one file - we'll only ever touch `en.js`
- No gulp!

This approach partially addresses the following:

- Automatically sort the langterms - not explicitly, but Serge returns them sorted it seems
- Easy to develop and test without requiring Serge - assuming it's existing langterms, yes

This explicitly _doesn't_ address the following:

- Language files per component - as far as I can tell, this is not actually possibly in conjuction with `import`, see below
- Github action to check langterms - not explicitly forbidden by this approach, just not addressed by it either.

Worth noting that this sort of undoes #448, and flattens FACE back down to one set of langterm files - this is because if `import` is handed anything other than an absolute path, bundling doesn't work correctly. Not sure what the best solution is here, but generally we have only one lang folder per component anyway, so maybe not a big deal.