Re-introduce multiple lang folders
While many projects do only use a single `lang` folder in their root, this is going to be a mess with FACE in the long term, as there will be plenty of domain-specific langterms that would all be mushed together. This re-introduces the lang folder specific to the assignment activity editor, but the caveat of this is that we aren't able to remove the [explicit inclusion of these lang folders](https://github.com/Brightspace/brightspace-integration/blob/39f2e6f71c34c5a6e2e47bbcd0426826623a3fc8/polymer.json#L29) from BSI.

The reason for this is that BSI bundles things, and `polymer build` is smart enough to include anything that is imported directly, but only if the `import` uses a literal value. Another option to this approach here would be to have a specific method for each component that calls into `localization.js`, but that would prevent us from ever making `localization.js` a generic module. In other words, pick two:

1. Having BSI not be aware of its dependencies' folder structures
2. Working with JS `import` rather than JSON and `fetch`
3. Making `localization.js` a generic module

This approach picks 2 and 3 as the most important.