Fix/reformat translations for FACE
These aren't a complete set, as we didn't get back a complete set from Serge last run. However, this _should_ be an accurate representation of what would have come back, if the current config was used.

Once the next run comes back, the gaps should be filled in here, and we'll be able to update `localization.js` to include all languages.