US111949 Add attachments editor
This adds the first phase of components that provide the ability to add/remove attachments associated with an assignment. As of this change, the `d2l-activity-attachments-editor` is quite simple and bare, and intentionally designed such that it could be used with the concept of activity attachments (as well the existing assignment attachments).

What this does do:

- Display an attachment-picker component in FACE
- Show action buttons with tooltips based on what types of (link) attachments are available to be added

What this doesn't do:

- Do stuff when you click on a button - coming soon, still figuring that out
- Do anything related to file attachments, as these are being handled separately (but should be compatible with these changes)

This was added as a new folder/Serge config since it's _probably_ generic enough that it could be moved up in the repo, outside of `d2l-activity-editor`, but I didn't want to go that far just yet. In particular, if we do decide to make file attachments an activities-wide concept, then moving this up might make sense. (Link attachments are still only supported in the assignments domain, but file attachments are supported elsewhere.)