US110586 Add picker dialogs
This adds handlers for the attachment picker buttons, which will call into the LMS' dialog JavaScript to open the correct picker dialog, and handle the result. In the FRA world, this behaviour is available within an ifrau service, but sadly there doesn't appear to be any such library/service that we can use in the non-FRA world, which means somewhat gross code with a lot of parameters. This largely mimics what the ifrau service is doing under the hood.

The behaviour here is as expected - clicking a button opens the dialog, hitting the Cancel button will do nothing, and selecting something will call the `/attach-link` route on the assignment folder to add the new attachment. This does not yet include listing the attachments, nor does it deal with file attachments at all.

A minor change to `save-status-mixin` here as well - by re-throwing the error, we allow for the caller to do things upon error (as well as the save-status component, which is listening for the `d2l-siren-entity-save-error` event).