US112016 Add attachment list
This adds a list view of (link) attachments in the editor, as well as support for removing them.

Thanks to the entity store and a small change in the LMS that adds `Link` headers to the add/delete attachment routes, this approach automatically re-renders the list when an attachment is added or removed without any additional effort. It's likely that once autosave is figured out a bit more, we'll have to add better interim states, but that remains to be seen.