US110586 Don't hide Insert Quicklink title
When `customTitle=''`, this hides the "Insert Quicklink" title on the dialog when it opens. However, there is a small bug somewhere in the LMS's dialog code such that when there is no title, the close button's position is slightly broken, so we'll just leave the title in, and only hide it for the URL dialog. (OneDrive and Google Drive use the `customTitleLangTerm`, which provides its own title and avoids the issue.)