US112016 Unfurling support
This adds support for attachments being able to unfurl themselves into rich previews via the AF unfurling service. This is done using the newish Requester/Provider approach, which allows the top-level `d2l-activity-editor` component to provide the configuration necesssary for the nested `d2l-labs-attachment` component via the `d2l-request-provider` event.

The behaviour that is currently being encompassed in `d2l-activity-editor`'s constructor (via `addEventListener`) will soon be provided via `BrightspaceUI/core`, but until that is merged and released, this is a simplified version of it that unblocks this change. This also relies on the LMS setting the `unfurlEndpoint` attribute on the `d2l-assignment-activity-editor` that it renders, which has now been merged.