US112016 Add trustedSitesEndpoint
This is a method of determining whether or not a given link attachment is a trusted site or not. Sadly, the actual Trusted Sites tool doesn't have its own API, but we can instead engage the HTML Editor API to determine if a link is trusted. The approach is a bit hackish, but it more-or-less mimics the behaviour provided by the `filterHtml` ifrau service.