Fix alignment of tooltip
The "Add link to existing activity" tooltip was being cut off, due to the boundary not being set. This applies the boundary to each tooltip - not strictly necessary, but this means that the order doesn't really matter. (If the styling were applied using CSS, we could `first-of-type`, but that doesn't allow for the computation that `d2l-tooltip` does with `boundary`.)