Fix HTML Editor in Edge
Sadly, the root cause here was weird DOM stuff by Edge, where the `d2l-activity-editor` somehow ends up as a child of the `d2l-activity-assignment-editor`, which is... odd. This means that properties were not being fully propagated through the elements.

Rather than adding the `htmlEditorEnabled` property to even more elements that don't actually care about it, this instead uses the new Provider/Requester approach to pass the property down to the `d2l-activity-text-editor` component, which is the only one that actually cares about it. This functionality has been added to BrightspaceUI/core, but not yet released - once it is, that will be used to replace most of this code.