US112016 Add untrusted URL alert
This uses `d2l-labs-attachment-list` to show a `d2l-alert` informing the user that they have attached an untrusted link when they do so. This also adds an `<li>` around `d2l-activity-attachment` as is being done in `assignments-ui`, as these are list elements, semantically speaking. The list is always in `editing` mode here, as FACE is an editor. This also required adding a sort of pass-through parameter, `creating`, to `d2l-activity-attachment` - this is because the `d2l-labs-attachment-list` looks for this parameter on the event target when deciding whether or not to show the untrusted URL alert.

A caveat to this change is that right now, the "Attachments" header shows above the `d2l-labs-attachment-list` when there are attachments, which we don't want to always show.