US112016 Set `creating` explicitly
This sets the `creating` value on the `d2l-activity-attachment` if and only if the user has actually added an attachment. This prevents the untrusted alert from showing all the time, even when the links were previously added.

This solution works, but also isn't very pretty - the event goes from the picker to the attachments-editor, then is handed back down the component chain via properties.