Allow for better partial re-rendering
With https://github.com/BrightspaceHypermediaComponents/siren-sdk/pull/109, we can now more precisely control partial re-renders of the page as things load. This change makes it so that when the page is first loading, we show the "Loading" text in place of the full editor. Once the page has finished loading once (which we know by `d2l-pending-resolved` being fired), we won't ever show that text again. This allows for lower-level components on the page to re-fetch data without the whole page looking like it's reloading.

This approach is also more compatible with lower-level `PendingContainer` elements - previously, when we were fetching a new attachment, the whole page would go to the "Loading" state, which meant we wouldn't even see the nice attachment skeleton that `d2l-labs-attachment` has!