Fix undefined hrefs
This was introduced in https://github.com/BrightspaceHypermediaComponents/activities/pull/545, or more accurately was surfaced by it. HTML attributes are strings, and LitElement properties are `undefined` by default - which, when stringified, is `"undefined"`. This can result is us trying to fetch weird URLs with "undefined".

To remedy this, it's a safe idea to initialize our strings in the constructor. Not required for everything, and not really needed for booleans (unless they should default to true), but a good practise to be in.