US112361 Add file selector dialog for attachments
Similar to change made in #500 for link attachments, this adds and connects a button to open a file picker dialog. Also similar to #500, this required getting deep into LMS dialog JavaScript, which is ugly and not very maintainable.

This approach assumes we are living on an LMS page that has this JS available - fine since that is currently the case, but not great. The code is also very finicky, with a bunch of mystery parameters and such. A better solution, which is currently in the works, is for us to have a `d2l-dialog` that supports a filepicker and use that - in the short term, it will likely still mean this code exists somewhere (within the `d2l-dialog`), but for our use case here, it will mean we're just interacting with a WC with a clean API.