Use QuickLinkTemplate to allow for copying
If we use the QuickLink directly, copying the assignment to another course will end up causing issues as the copied assignment will have attachments in another course. Instead, we can use the QuickLinkTemplate so that when an assignment is copied to another course, the attachments will refer to the new org unit.