Fix flash of generic link icon
Issue: when adding a new link, there is a brief flash of the generic-link icon, before switching to the attachment skeleton, then finally the unfurled link. This is because of a mismatch in behaviour between `d2l-activity-attachment` and `d2l-labs-attachment` - the former expects the latter to be able to handle null attributes/inputs, and the latter expects the former to never provide null attributes/inputs!

This fix changes `d2l-activity-attachment` to never pass a null `attachment` to the `d2l-labs-attachment` component, instead opting to just not render anything. This is probably a good approach to take in general, actually - it results in a faster (no-op) render when the component is first created.