US111908 Refactor dialog openers
This change is possible due to a server-side change that makes the functions to open the various dialogs available via the `ObjectRepository` on the page. This still relies on magic keys to retrieve these opener functions, but this removes the majority of the deep-LMS-knowledge code from the UI side, which is an improvement.

This is functionally identical to what already existed.