Allow tabbie to calculate its position
There is JavaScript in the LMS that calculates where to place the opt out tabbie so that it sits under the immersive nav bar. When FACE is loading, the editor (including the immersive nav) is hidden and has a height of 0. Thus, it is unable to calculate where to place the opt-out tabbie. This commit changes it so that the loading screen becomes a fixed overlay over the editor and the editor compoonents render beneath it.