Set original gradeHref for scoreAndGrade
We cannot rely on the `scoreAndGrade.associatedGrade` value initially as the initial value is null. It is null because i needs to be a grade candidate, but the GradeCandidateCollection store is a separate thing that is not loaded yet. So I added a `gradeHref` which can be initialized based on the grade link that would be present on the ActivityUsage entity if the activity is associated to a grade.