Fix undefined selected grade candidate error
Throws when there are no grade candidates. We're actually not going to be rendering the dropdown when there are no grade candidates, so the code that is throwing these errors won't be executed in the future and so these errors won't happen. But it's probably safer to check if `selected` is undefined in case there are future changes to this that could potentially render the dropdown.