Remove linked grade didn't change check.
The code being removed checked if the linked grade didn't change when we clicked 'OK' in the dialog, and if it didn't change, we would not change the assignment `out of` value in the input. We are removing this check so that if the user doesn't change the `link to existing grade` and click 'OK', we will still override whatever assignment `outOf` value the user has entered.