Fix MobX error - changing observed observable values outside actions is not allowed
Loading special access is an async function. Loading Competencies or Outcomes after the async function was causing the MobX error as it would be changing observed values after another async.
I've changed it so that the two async functions to load special access and load competencies or outcomes are in a Promise.all.