US117525 - Fetch updated special-access info after dialog closes
Initially, we expected the data to be returned when the dialog closed, so we could skip doing a fetch on `/special-access`. Although this approach worked, it required a lot of changes to the back-end. So to keep it simple, we will do a fetch after the dialog closes and the dialog will not be expected to return any info.