Rename misleading `isReadOnly` Assignment MobX property
The `isReadOnly` MobX assignment property was specific to whether the Assignment Type was editable or not. So when there are no groups, the assignment type is not editable (i.e. cannot switch between individual and group assignment type). The `isReadOnly` property would be `true` even though the assignment itself is editable, just that the assignment type property is not. I found this naming confusing, so I've renamed the property to be specific to assignment type editability.