DE39814 - Allow an `order` property to be set for each editor
This order property will be used to determine the order of validation and saving when the save event is triggered.
The default order will be 1000.
This PR sets the activity-assignment-editor to have an order of 2000 and so the assignment PATCH/save will occur after the other activity-editor APIs.

The problem we were having was that there was a corrupted assignment that was associated to a grade, but had no `score-out-of` value. Even when we fixed the `score-out-of` value, the assignment failed to save, because the assignment PATCH request occurred before the `score-out-of`POST, and would throw up because it saw that the assignment was associated to a grade but had no score. So this ordering fix allows us to call the `score-out-of` API to update the assignment score with a value before sending the assignment PATCH request. And now when the assignment saves, it won't throw up because there is now a valid out of score on the assignment.