US123692 - changes to support working copy dirty check
The component calling the dirty check is the `quiz-editor-detail` component. It will check that the base quiz is not dirty, then check that it's checked out working copy does not have the `canCheckin` action. It avoids an infinite recursion by not passing in the `quizStore` the second time, so that it can't fetch the working copy's working copy.