Improve timing summarizer update performance
Improve timing summarizer update performance after a dialog check in by temporarily replacing the checked out timing entity with the entity from the dialog so that the UI updates immediately. We still fetch the checked out timing entity again, just to make sure the summarizer info we are displaying matches the working copy state for timing.