Fix incorrect existing grade displayed after saving
In `activity-score-grade` `load` function, we set `gradeCandidateCollection` and `newGradeCandidatesCollection` to `null` so that when we open the "choose from grades" dialog for the first time (or after a save), it will know to re-fetch the collections. The problem is, the collection entities are still in the store, so they don't actually end up being re-fetched and we end up displaying a stale incorrect linked grade item after a save.
This also fixes a problem where when we "Create new grade" and "save", the "gradeCandidateCollection" wasn't being re-fetched, so the grade we just created and linked to does not appear in the link to existing dropdown even though the assignment is linked to it.

https://trello.com/c/IPhZb0hO/5-choose-grades-incorrect-selected-existing-grade-item-showing-after-save