This switches grades dialog to use ActivityEditorWorkingCopyDialogMixin
This PR switches the grades dialog from ActivityEditorMixin to ActivityEditorWorkingCopyDialogMixin which has a bunch of helpers to help open/close the dialog and checkout/checkin.
The PR also simplifies the showing of the "+ Grade Category" button a little.
There are 2 issues with this PR. 1: the `async` state on the dialog should be `true` to show the loading spinner. 2: when the selectbox LD flag is off, it is still checking out because the entity contains a checkout action, although it doesn't patch or checkin or do anything with the checked out entity. But I should prevent that somehow probably.