Simplify some of the hrefs referenced in the grades dialog
In the `d2l-activity-grades-dialog` file,
for non-working copy `this.baseActivityUsageHref`, `this.href`. and `this.checkedOutHref` are the same. `this.dialogHref` should not exist.
For working copy, `this.baseActivityUsageHref` is different from `this.href`. `this.href` is the same as `this.checkedOutHref`. `this.dialogHref` is defined and different from the other 2 hrefs.