Fix dialog height and async
Non working copy (i.e. createSelectboxGradeItemEnabled is false) will not be using the async loading spinner in the dialog. It will open when it is ready.
Working copy (i.e. createSelectboxGradeItemEnabled is true) will be using the async loading spinner in the dialog. The loading spinner still acts a bit funny and might disappear too soon before everything is loaded.