Fix cannot re-open dialog after an 'x' click to close dialog
This is due to it initially only calling `_cancel` on cancel button click. But that function also needs to be called if the dialog is closed via the 'x' or 'esc'. The only way to catch this is to listen to the `d2l-dialog-close` event. The change to not pass the event to our dialog mixin is because the mixin is doing `event.stopPropagation` which was preventing the `done` action in `ok` dialog button presses from being present in the `d2l-dialog-close` event.