Fix issue of individual grades/categories/schemes being fetched individually
I've moved the fetching of grades, categories, and schemes to only be fetched when the dialog is opened. Otherwise, we are doing unnecessary fetching.
This fixes the issue of it fetching each individual grade, category, and scheme when going from ungraded to in grades due to us not fetching it unless we open the dialog.

However, going from Not In Grades to opening the dialog still fetches each individual item, why? There is a problem when we switch gradebookStatus immediately before we fetch those items. I don't know why. Is it an issue with the store and how both the gradebookStatus update and the grade/categories/schemes fetch happening too close together or something? Because it is fine if we go from ungraded to in grades (which causes a gradebookStatus update), then open the dialog, which no longer does a gradebookStatus update, but will fetch the items but not each individual one as it respects the cache-primer.